+++ CHANGELOG.md
@@ -3,13 +3,16 @@
 Notable changes to  Format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). Project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!--
+## [Unreleased]
+
 -->
 
-## [Unreleased]
+## [0.4.1] - 20240806
 
 ### Added
 
 - demo `demo/zipcode.py`
+- support for multiple workers & CORS headers (`--workers` & `--cors_origin` cmdline option)
 
 ### Fixed